+++ src/main/java/org/riekr/jloga/react/Observable.java
@@ -1,6 +1,10 @@
 package org.riekr.jloga.react;
 
+import java.awt.*;
+import java.awt.event.HierarchyEvent;
+import java.awt.event.HierarchyListener;
 import java.util.concurrent.Future;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
@@ -10,6 +14,30 @@ public interface Observable<T> {
 
 	@NotNull
 	Unsubscribable subscribe(Observer<? super T> observer);
+
+	@NotNull
+	default Unsubscribable subscribe(Component component, Observer<? super T> observer) {
+		AtomicReference<Unsubscribable> unsubscribable = new AtomicReference<>(subscribe(observer));
+		HierarchyListener hierarchyListener = (e) -> {
+			if ((e.getChangeFlags() & HierarchyEvent.PARENT_CHANGED) != 0) {
+				if (component.getParent() == null) {
+					Unsubscribable u = unsubscribable.getAndSet(null);
+					if (u != null)
+						u.unsubscribe();
+				} else {
+					if (unsubscribable.get() == null)
+						unsubscribable.set(subscribe(observer));
+				}
+			}
+		};
+		component.addHierarchyListener(hierarchyListener);
+		return () -> {
+			Unsubscribable u = unsubscribable.get();
+			if (u != null)
+				u.unsubscribe();
+			component.removeHierarchyListener(hierarchyListener);
+		};
+	}
 
 	@NotNull
 	default Unsubscribable subscribe(Consumer<? super T> onNext, Consumer<Throwable> onError) {